utils/pyproject.toml
[tool.poetry]
name = "utils"
version = "0.1.0"
description = ""
authors = ["Danut Codrescu <danutzcodrescu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
aiohttp = "^3.7.4"
asyncio = "^3.4.3"
python-dotenv = {git = "git@github.com:theskumar/python-dotenv.git"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
